refactor author list into component
