+++ src/templates/post.tsx
@@ -3,7 +3,7 @@ import { graphql, Link } from 'gatsby';
 import Img, { FluidObject } from 'gatsby-image';
 import * as _ from 'lodash';
 import { lighten, setLightness } from 'polished';
-import React, { useState } from 'react';
+import React from 'react';
 import { Helmet } from 'react-helmet';
 
 import { css } from '@emotion/core';
@@ -11,7 +11,6 @@ import styled from '@emotion/styled';
 
 import { Footer } from '../components/Footer';
 import SiteNav, { SiteNavMain } from '../components/header/SiteNav';
-import { AuthorList, AuthorListItem, AuthorProfileImage } from '../components/PostCard';
 import PostContent from '../components/PostContent';
 import { ReadNext } from '../components/ReadNext';
 import { Subscribe } from '../components/subscribe/Subscribe';
@@ -20,6 +19,17 @@ import IndexLayout from '../layouts';
 import { colors } from '../styles/colors';
 import { inner, outer, SiteMain } from '../styles/shared';
 import config from '../website-config';
+import { AuthorList } from '../components/AuthorList';
+
+export interface Author {
+  id: string;
+  bio: string;
+  avatar: {
+    children: Array<{
+      fluid: FluidObject;
+    }>;
+  };
+}
 
 interface PageTemplateProps {
   pathContext: {
@@ -47,15 +57,7 @@ interface PageTemplateProps {
         };
         excerpt: string;
         tags: string[];
-        author: Array<{
-          id: string;
-          bio: string;
-          avatar: {
-            children: Array<{
-              fluid: FluidObject;
-            }>;
-          };
-        }>;
+        author: Author[];
       };
     };
     relatedPosts: {
@@ -89,7 +91,7 @@ export interface PageContext {
   frontmatter: {
     image: {
       childImageSharp: {
-        fluid: any;
+        fluid: FluidObject;
       };
     };
     excerpt: string;
@@ -97,15 +99,7 @@ export interface PageContext {
     date: string;
     draft?: boolean;
     tags: string[];
-    author: Array<{
-      id: string;
-      bio: string;
-      avatar: {
-        children: Array<{
-          fluid: FluidObject;
-        }>;
-      };
-    }>;
+    author: Author[];
   };
 }
 
@@ -206,71 +200,7 @@ const PageTemplate: React.FC<PageTemplateProps> = props => {
                 </PostFullCustomExcerpt>
                 <PostFullByline className="post-full-byline">
                   <section className="post-full-byline-content">
-                    <AuthorList className="author-list">
-                      {post.frontmatter.author.map(author => {
-                        const [hovered, setHover] = useState(false);
-                        let timeout: NodeJS.Timeout;
-                        function handleMouseEnter() {
-                          clearTimeout(timeout);
-                          setHover(true);
-                        }
-
-                        function handleMouseLeave() {
-                          clearTimeout(timeout);
-                          timeout = setTimeout(() => setHover(false), 600);
-                        }
-
-                        return (
-                          <AuthorListItem
-                            key={author.id}
-                            className="author-list-item"
-                            onMouseEnter={() => handleMouseEnter()}
-                            onMouseLeave={() => handleMouseLeave()}
-                          >
-                            <div
-                              css={css`
-                                ${AuthorCardStyle} ${hovered && Hovered}
-                              `}
-                              className="author-card"
-                            >
-                              {author.avatar.children.length && (
-                                <Img
-                                  css={AuthorProfileImage}
-                                  className="author-profile-image"
-                                  fluid={author.avatar.children[0].fluid}
-                                  fadeIn={false}
-                                />
-                              )}
-                              <div className="author-info">
-                                <div className="bio">
-                                  <h2>{author.id}</h2>
-                                  <p>{author.bio}</p>
-                                  <p>
-                                    <Link to={`/author/${_.kebabCase(author.id)}/`}>
-                                      More posts
-                                    </Link>{' '}
-                                    by {author.id}.
-                                  </p>
-                                </div>
-                              </div>
-                            </div>
-                            <Link
-                              css={AuthorAvatar}
-                              className="author-avatar"
-                              to={`/author/${_.kebabCase(author.id)}/`}
-                            >
-                              <Img
-                                css={AuthorProfileImage}
-                                className="author-profile-image"
-                                fluid={author.avatar.children[0].fluid}
-                                alt={author.id}
-                                fadeIn={false}
-                              />
-                            </Link>
-                          </AuthorListItem>
-                        );
-                      })}
-                    </AuthorList>
+                    <AuthorList authors={post.frontmatter.author} tooltip="large" />
                     <section className="post-full-byline-meta">
                       <h4 className="author-name">
                         {post.frontmatter.author.map(author => (
@@ -468,127 +398,6 @@ const PostFullByline = styled.div`
       color: #fff;
     }
   }
-`;
-
-export const AuthorAvatar = css`
-  display: block;
-  overflow: hidden;
-  margin: 0 -4px;
-  width: 40px;
-  height: 40px;
-  border: #fff 2px solid;
-  border-radius: 100%;
-  transition: all 0.5s cubic-bezier(0.4, 0.01, 0.165, 0.99) 700ms;
-
-  @media (max-width: 500px) {
-    width: 36px;
-    height: 36px;
-  }
-
-  @media (prefers-color-scheme: dark) {
-    /* border-color: color(var(--darkgrey) l(+2%)); */
-    border-color: ${lighten('0.02', colors.darkgrey)};
-  }
-`;
-
-const AuthorCardStyle = css`
-  position: absolute;
-  bottom: 130%;
-  left: 50%;
-  z-index: 600;
-  display: flex;
-  justify-content: space-between;
-  margin-left: -200px;
-  width: 400px;
-  font-size: 1.4rem;
-  line-height: 1.5em;
-  background: white;
-  border-radius: 3px;
-  box-shadow: rgba(39, 44, 49, 0.08) 0 12px 26px, rgba(39, 44, 49, 0.06) 1px 3px 8px;
-  opacity: 0;
-  transition: all 0.35s cubic-bezier(0.4, 0.01, 0.165, 0.99);
-  transform: scale(0.98) translateY(15px);
-  pointer-events: none;
-  padding: 20px 20px 22px;
-
-  :before {
-    content: '';
-    position: absolute;
-    top: 100%;
-    left: 50%;
-    display: block;
-    margin-left: -8px;
-    width: 0;
-    height: 0;
-    border-top: 8px solid #fff;
-    border-right: 8px solid transparent;
-    border-left: 8px solid transparent;
-  }
-
-  .author-info {
-    flex: 1 1 auto;
-    padding: 0 0 0 20px;
-  }
-
-  .author-info h2 {
-    margin: 8px 0 0;
-    font-size: 1.6rem;
-  }
-
-  .author-info p {
-    margin: 4px 0 0;
-    color: color(var(--midgrey) l(-10%));
-  }
-
-  .author-info .bio h2 {
-    margin-top: 0;
-  }
-
-  .author-info .bio p {
-    margin-top: 0.8em;
-  }
-
-  .author-profile-image {
-    flex: 0 0 60px;
-    margin: 0;
-    width: 60px;
-    height: 60px;
-    border: none;
-  }
-
-  @media (max-width: 1170px) {
-    margin-left: -50px;
-    width: 430px;
-
-    :before {
-      left: 50px;
-    }
-  }
-
-  @media (max-width: 650px) {
-    display: none;
-  }
-
-  @media (prefers-color-scheme: dark) {
-    /* background: color(var(--darkmode) l(+4%)); */
-    background: ${lighten('0.04', colors.darkmode)};
-    box-shadow: 0 12px 26px rgba(0, 0, 0, 0.4);
-
-    :before {
-      /* border-top-color: color(var(--darkmode) l(+4%)); */
-      border-top-color: ${lighten('0.04', colors.darkmode)};
-    }
-  }
-`;
-
-const Hovered = css`
-  opacity: 1;
-  transform: scale(1) translateY(0px);
-  pointer-events: auto;
-`;
-
-const PostFullMetaDate = styled.time`
-  color: ${colors.blue};
 `;
 
 export const PostFullTitle = styled.h1`
